Backend done :)

+++ backend/controllers/goal.controller.js
@@ -1,10 +1,17 @@
 const asyncHandler = require("express-async-handler");
-const Goal = require('../models/goal.model')
+const Goal = require("../models/goal.model");
+const User = require("../models/user.model");
+
 // @decs    Get goals
 // @route   GET /api/goals
 // @access  private
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find()
+  const goals = await Goal.find({ user: req.user.id });
+  if (!goals) {
+    res.json({
+      message: "No Goals found",
+    });
+  }
   res.status(200).json(goals);
 });
 
@@ -18,8 +25,9 @@ const setGoal = asyncHandler(async (req, res) => {
   }
 
   const goal = await Goal.create({
-    text: req.body.text
-  })
+    text: req.body.text,
+    user: req.user.id,
+  });
 
   res.status(200).json(goal);
 });
@@ -28,14 +36,31 @@ const setGoal = asyncHandler(async (req, res) => {
 // @route   PUT /api/goals/:id
 // @access  private
 const updateGoal = asyncHandler(async (req, res) => {
-  const goal = Goal.findById(req.params.id);
-  if(!goal){
+  const goal = await Goal.findById(req.params.id);
+  if (!goal) {
     res.status(400);
     throw new Error("Goal not found");
   }
-  const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {new: true});
+
+  const user = await User.findById(req.user.id);
+
+  // Check for a user
+  if (!user) {
+    res.status(401);
+    throw new Error("USer not found");
+  }
+
+  //  Make the login user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
+
+  const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
+    new: true,
+  });
   res.status(200).json({
-    updatedGoal
+    updatedGoal,
   });
 });
 
@@ -43,14 +68,29 @@ const updateGoal = asyncHandler(async (req, res) => {
 // @route   DELETE /api/goals/:id
 // @access  private
 const deleteGoal = asyncHandler(async (req, res) => {
-  const goal = Goal.findById(req.params.id);
-  if(!goal){
+  const goal = await Goal.findById(req.params.id);
+  if (!goal) {
     res.status(400);
     throw new Error("Goal not found");
   }
-  const deletedGoal = await Goal.findByIdAndDelete(req.params.id);
+
+  const user = await User.findById(req.user.id);
+
+  // Check for a user
+  if (!user) {
+    res.status(401);
+    throw new Error("USer not found");
+  }
+
+  //  Make the login user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
+
+  await Goal.findByIdAndDelete(req.params.id);
   res.status(200).json({
-    id: req.params.id
+    id: req.params.id,
   });
 });
